# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saree Blouse Fitter API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Analyze blouse photo with AI
 */
export const AnalyzeBlouseBody = zod.object({
  imageBase64: zod.string(),
  userId: zod.string(),
});

export const AnalyzeBlouseResponse = zod.object({
  imageType: zod.enum(["person", "blouse"]),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string(),
  aiAnalysis: zod.string(),
  analysisPoints: zod.array(
    zod.object({
      label: zod.string(),
      detail: zod.string(),
    }),
  ),
  suggestedStyles: zod.array(zod.string()),
});

/**
 * @summary Get all blouse fits for a user
 */
export const GetBlouseFitsQueryParams = zod.object({
  userId: zod.coerce.string(),
});

export const getBlouseFitsResponseFindMyTailorDefault = false;

export const GetBlouseFitsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string().nullish(),
  stylePrefs: zod
    .object({
      neckline: zod.string().optional(),
      sleeves: zod.string().optional(),
      back: zod.string().optional(),
      fabric: zod.string().optional(),
      fit: zod.string().optional(),
    })
    .nullish(),
  aiAnalysis: zod.string().nullish(),
  notes: zod.string().nullish(),
  findMyTailor: zod.boolean().default(getBlouseFitsResponseFindMyTailorDefault),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const GetBlouseFitsResponse = zod.array(GetBlouseFitsResponseItem);

/**
 * @summary Save a blouse fit
 */
export const SaveBlouseFitBody = zod.object({
  userId: zod.string(),
  imageUrl: zod.string().nullish(),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string().nullish(),
  stylePrefs: zod
    .object({
      neckline: zod.string().optional(),
      sleeves: zod.string().optional(),
      back: zod.string().optional(),
      fabric: zod.string().optional(),
      fit: zod.string().optional(),
    })
    .nullish(),
  aiAnalysis: zod.string().nullish(),
});

/**
 * @summary Get a single blouse fit
 */
export const GetBlouseFitParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBlouseFitQueryParams = zod.object({
  userId: zod.coerce.string(),
});

export const getBlouseFitResponseFindMyTailorDefault = false;

export const GetBlouseFitResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string().nullish(),
  stylePrefs: zod
    .object({
      neckline: zod.string().optional(),
      sleeves: zod.string().optional(),
      back: zod.string().optional(),
      fabric: zod.string().optional(),
      fit: zod.string().optional(),
    })
    .nullish(),
  aiAnalysis: zod.string().nullish(),
  notes: zod.string().nullish(),
  findMyTailor: zod.boolean().default(getBlouseFitResponseFindMyTailorDefault),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a blouse fit
 */
export const DeleteBlouseFitParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteBlouseFitResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Submit an owned fit to the tailor list
 */
export const SubmitFitToTailorsParams = zod.object({
  id: zod.coerce.number(),
});

export const SubmitFitToTailorsBody = zod.object({
  userId: zod.string(),
});

export const submitFitToTailorsResponseFindMyTailorDefault = false;

export const SubmitFitToTailorsResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string().nullish(),
  stylePrefs: zod
    .object({
      neckline: zod.string().optional(),
      sleeves: zod.string().optional(),
      back: zod.string().optional(),
      fabric: zod.string().optional(),
      fit: zod.string().optional(),
    })
    .nullish(),
  aiAnalysis: zod.string().nullish(),
  notes: zod.string().nullish(),
  findMyTailor: zod
    .boolean()
    .default(submitFitToTailorsResponseFindMyTailorDefault),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update tailor notes for a fit
 */
export const UpdateFitNotesParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateFitNotesBody = zod.object({
  notes: zod.string(),
});

export const updateFitNotesResponseFindMyTailorDefault = false;

export const UpdateFitNotesResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string().nullish(),
  stylePrefs: zod
    .object({
      neckline: zod.string().optional(),
      sleeves: zod.string().optional(),
      back: zod.string().optional(),
      fabric: zod.string().optional(),
      fit: zod.string().optional(),
    })
    .nullish(),
  aiAnalysis: zod.string().nullish(),
  notes: zod.string().nullish(),
  findMyTailor: zod
    .boolean()
    .default(updateFitNotesResponseFindMyTailorDefault),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Get fits submitted to the tailor list
 */
export const GetTailorCustomersQueryParams = zod.object({
  tailorId: zod.coerce.string(),
});

export const getTailorCustomersResponseFindMyTailorDefault = false;

export const GetTailorCustomersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  imageUrl: zod.string().nullish(),
  thumbnailUrl: zod.string().nullish(),
  measurements: zod
    .object({
      bust: zod.number().optional(),
      waist: zod.number().optional(),
      shoulder: zod.number().optional(),
      hip: zod.number().optional(),
    })
    .nullish(),
  bodyShape: zod.string().nullish(),
  stylePrefs: zod
    .object({
      neckline: zod.string().optional(),
      sleeves: zod.string().optional(),
      back: zod.string().optional(),
      fabric: zod.string().optional(),
      fit: zod.string().optional(),
    })
    .nullish(),
  aiAnalysis: zod.string().nullish(),
  notes: zod.string().nullish(),
  findMyTailor: zod
    .boolean()
    .default(getTailorCustomersResponseFindMyTailorDefault),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const GetTailorCustomersResponse = zod.array(
  GetTailorCustomersResponseItem,
);