# lib/api-zod/src/generated/types/analyzeBlouseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saree Blouse Fitter API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeBlouseRequest {
  imageBase64: string;
  userId: string;
}