# lib/api-zod/src/generated/types/analyzeBlouseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saree Blouse Fitter API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisPoint } from "./analysisPoint";
import type { AnalyzeBlouseResponseImageType } from "./analyzeBlouseResponseImageType";
import type { Measurements } from "./measurements";

export interface AnalyzeBlouseResponse {
  imageType: AnalyzeBlouseResponseImageType;
  measurements?: Measurements | null;
  bodyShape: string;
  aiAnalysis: string;
  analysisPoints: AnalysisPoint[];
  suggestedStyles: string[];
}