# lib/api-zod/src/generated/types/measurements.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saree Blouse Fitter API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Measurements {
  bust?: number;
  waist?: number;
  shoulder?: number;
  hip?: number;
}